fix(seo): keep wallet SDK out of SSR; honor profile search opt-out
- Extract the Phantom-SDK-backed provider into wallet-phantom.tsx, loaded
  via a no-SSR dynamic import; @phantom/react-sdk is no longer in the
  server runtime bundle while the wallet context still server-renders so
  page content reaches the initial HTML.
- Derive profile robots.index from profileVisibility.searchEngineIndexing
  so search-indexing opt-outs are honored (base + tab routes); document
  why the sitemap relies on the page-level robots signal.
